refactor(memory,cli): pure namespace/reconcile subpaths keep node:sqlite off the request path
@dawn-ai/memory's barrel re-exports sqliteMemoryStore, so importing a
namespace helper from it dragged node:sqlite into every consumer. Add
"./namespace" and "./reconcile" subpath exports (both zero-import modules) and
repoint the runtime's four call sites at them. buildMemoryContext moves to a
new node-free memory-context.ts; resolve-memory.ts (which reads dawn.config.ts
and lazily opens the sqlite store) re-exports it so callers are unchanged.
The memory-approve HTTP route now takes resolveIdentityKeys as an injected
function, falling back to the default semantic identity when the runtime has
no filesystem to read memory.ts from.
